Starting to store and autofill data.
At this point, the code demonstrates a simple router, which serves
as a primary switch for the application. All functionality is broken
out into functions to handle the saving of data, and forms are
templated with autofill where appropriate.

One issue here is that we get the username, which is not necessarily
a unique identifier, but we trust the BNumber more. We probably have
to trust the username, because that is the authentication token. There
is no way to authenticate a BNumber... it's a bogus piece of information
that the user provides.

The alternative is that we take the username and look up the Bnumber...
but, at the end of the day, the thing the user gives us through
their authetication is a username... which is probably what we should
consider as a 'primary key.'

To be debated.