Updated to handle new model, fixed start()
This involves some changes and learning as I updated the data model.

The model has more foreign key constraints now, and a linking table.

For example:

    facQ = (Faculty.select()
      .join(FacultyProjects, on = (FacultyProjects.fid == Faculty.fid))
      .join(Projects, on = (FacultyProjects.pid == Projects.pid))
      .where(Faculty.username == username)
      .where(FacultyProjects.corresponding == True)
      )

This is the join to pull out the faculty member if they are a corresponding
author on one project.

projectQ = (Projects.select()
        .join(Faculty, on = (theFaculty.fid == Projects.corresponding))
        )

pulls out all the projects which that faculty member is a corresponding member
for... and there should be only one. I'm not sure if we can enforce that,
except for making sure that one faculty member cannot insert more than one
project. Hm. That is probably not handled at the moment.

There's also a bunch of "session timeout" code that is almost entirely
untested, but is something that we should explore/discuss further. The session
(read: cookie) information is important, possibly, and I would like to make
sure I know how/why I'm using it. It isn't clear how it is valuable.