Moved code to JavaScript
I was doing a lot of work to do DB queries to populate web pages with template
variables. For example, on the start page, I wanted to look into the DB and
pre-load the fields if someone had already visited. This required a lot of work
in Python. The alternative: create a new endpoint in Flask that handles a JSON
request.

This approach lets me set id = "modelField" in my HTML for all of the text
fields. If the fields have the same id's as the names of the model fields, then
I can do a JSON request, get back an object, and walk the keys of that object.
These can automatically be turned into IDs on the webpage, and I can populate
the values of the form (if the values existed).

That may have made no sense. However, the end-result is less code on the server
(or, a simplified interface...), because now I've moved all form population
based on old data into a GET endpoint (or a POST... I should think about that),
and that endpoint does a query and simply JSONifies the object that is found.
That object is walked, and the values in the object directly stuffed into
fields.

Honest. It's easier.