Adding in documentation.
This is written with mkdocs, which is a cute little Markdown-based
documentation tool.

Some guidelines. If we need to write more docs (we do), then this is a good
place to do it.

mkdocs serve

from within the docs/urcpp directory to run a local server (port 8000) where
you can view the documentation that is written.