Working collaborators page.
Both the Python and JavaScript for this page are a bit ugly.

Specifically, I am certain we can simplify the SQL work. I have a series
of sequential operations that are probably made simpler with good
queries. It is clear, but could be better. Also, we probably
have some queries that can be pulled out as functions for re-usability.

On the JavaScript side, I found things got long as well. Specifically,
there's a lot of JavaScript to handle the adding/removing of fields,
keeping track of indicies, and making sure the state of the BNumber fields
is correct. This involves doing checks against the server
pretty rapidly... we'll see how that goes in practice.